Migration du namespace GeneralMeeting vers AssembleeGenerale
