+++ sources/AppBundle/Controller/Admin/Members/GeneralMeetingQuestion/EditAction.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace AppBundle\Controller\Admin\Members\GeneralMeetingQuestion;
 
 use AppBundle\AssembleeGenerale\Entity\Repository\QuestionRepository;
-use AppBundle\AssembleeGenerale\Form\GeneralMeetingQuestionFormType;
+use AppBundle\AssembleeGenerale\Form\QuestionFormType;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
@@ -26,7 +26,7 @@ class EditAction extends AbstractController
             throw $this->createAccessDeniedException('Seules les questions en attente peuvent être modifiées');
         }
 
-        $form = $this->createForm(GeneralMeetingQuestionFormType::class, $question);
+        $form = $this->createForm(QuestionFormType::class, $question);
         $form->handleRequest($request);
         if ($form->isSubmitted() && $form->isValid()) {
             $this->questionRepository->save($question);
