+++ sources/AppBundle/Controller/Admin/Members/GeneralMeetingVote/ListAction.php
@@ -6,7 +6,7 @@ namespace AppBundle\Controller\Admin\Members\GeneralMeetingVote;
 
 use AppBundle\AssembleeGenerale\Entity\Repository\QuestionRepository;
 use AppBundle\AssembleeGenerale\Entity\Repository\VoteRepository;
-use AppBundle\GeneralMeeting\GeneralMeetingRepository;
+use AppBundle\AssembleeGenerale\Entity\Repository\PresenceRepository;
 use DateTimeImmutable;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
@@ -15,7 +15,7 @@ use Twig\Environment;
 class ListAction
 {
     public function __construct(
-        private readonly GeneralMeetingRepository $generalMeetingRepository,
+        private readonly PresenceRepository $generalMeetingRepository,
         private readonly QuestionRepository $questionRepository,
         private readonly VoteRepository $voteRepository,
         private readonly Environment $twig,
